setup login page and landing page
