+++ components/NavHome.vue
@@ -1,0 +1,13 @@
+<template>
+    <nav class="text-white border-none flex bg-primary justify-between items-center shadow-md mx-auto h-18 py-1">
+        <div class="ml-20 flex flex-col items-center">
+            <img src="/assets/images/GraduationCap.png" class="w-8">
+            <p class="font-bold">SchoolPilot</p>
+        </div>
+        <div class="mr-10 space-x-6">
+            <NuxtLink class="font-medium" to="/loginPage">Login</NuxtLink>
+            <NuxtLink class="font-medium" to="/signup">Signup</NuxtLink>
+        </div>
+    </nav>
+
+</template>
