+++ layouts/home.vue
@@ -1,0 +1,4 @@
+<template>
+    <NavHome></NavHome>
+    <slot/>
+</template>
